qcacld-3.0: Fix memory corruption of GET_PREFERRED_FREQ_LIST
1. Change pcl_list to uint32_t array.
2. Populate weight freq from pcl_list.
3. Correct the skb allocation size to include attr hdr

CRs-Fixed: 2554031

+++ components/cmn_services/policy_mgr/src/wlan_policy_mgr_pcl.c
@@ -312,7 +312,7 @@ static QDF_STATUS policy_mgr_modify_pcl_based_on_dnbs(
 						uint32_t *pcl_len_org)
 {
 	uint32_t i, pcl_len = 0;
-	uint8_t pcl_list[QDF_MAX_NUM_CHAN];
+	uint32_t pcl_list[QDF_MAX_NUM_CHAN];
 	uint8_t weight_list[QDF_MAX_NUM_CHAN];
 	bool ok;
 	QDF_STATUS status = QDF_STATUS_E_FAILURE;
